+++ src/index/author_time.go
@@ -18,13 +18,13 @@ func NewAuthorTimeIndex(idx Index, kb KeyBuilder) *AuthorTimeIndex {
 }
 
 // InsertEvent inserts an entry keyed by (pubkey, kind, created_at).
-func (a *AuthorTimeIndex) InsertEvent(ctx context.Context, pubkey [32]byte, kind uint32, createdAt uint64, loc types.RecordLocation) error {
+func (a *AuthorTimeIndex) InsertEvent(ctx context.Context, pubkey [32]byte, kind uint16, createdAt uint32, loc types.RecordLocation) error {
 	key := a.kb.BuildAuthorTimeKey(pubkey, kind, createdAt)
 	return a.idx.Insert(ctx, key, loc)
 }
 
 // RangeByAuthor returns entries for an author and kind between time range.
-func (a *AuthorTimeIndex) RangeByAuthor(ctx context.Context, pubkey [32]byte, kind uint32, from uint64, to uint64, desc bool) (Iterator, error) {
+func (a *AuthorTimeIndex) RangeByAuthor(ctx context.Context, pubkey [32]byte, kind uint16, from uint32, to uint32, desc bool) (Iterator, error) {
 	minKey := a.kb.BuildAuthorTimeKey(pubkey, kind, from)
 	maxKey := a.kb.BuildAuthorTimeKey(pubkey, kind, to)
 	if desc {
