+++ src/index/search.go
@@ -18,13 +18,13 @@ func NewSearchIndex(idx Index, kb KeyBuilder) *SearchIndex {
 }
 
 // InsertSearchEntry inserts a search index entry.
-func (s *SearchIndex) InsertSearchEntry(ctx context.Context, kind uint32, searchType SearchType, tagValue []byte, createdAt uint64, loc types.RecordLocation) error {
+func (s *SearchIndex) InsertSearchEntry(ctx context.Context, kind uint16, searchType SearchType, tagValue []byte, createdAt uint32, loc types.RecordLocation) error {
 	key := s.kb.BuildSearchKey(kind, searchType, tagValue, createdAt)
 	return s.idx.Insert(ctx, key, loc)
 }
 
 // RangeSearch returns search entries for the given prefix.
-func (s *SearchIndex) RangeSearch(ctx context.Context, kind uint32, searchType SearchType, tagValuePrefix []byte, desc bool) (Iterator, error) {
+func (s *SearchIndex) RangeSearch(ctx context.Context, kind uint16, searchType SearchType, tagValuePrefix []byte, desc bool) (Iterator, error) {
 	minKey, maxKey := s.kb.BuildSearchKeyRange(kind, searchType, tagValuePrefix)
 	if desc {
 		return s.idx.RangeDesc(ctx, minKey, maxKey)
