+++ src/storage/storage_test.go
@@ -257,7 +257,7 @@ func TestScannerSinglePage(t *testing.T) {
 		event := &types.Event{
 			ID:        [32]byte{byte(i)},
 			Pubkey:    [32]byte{0x01},
-			CreatedAt: uint64(1655000000 + i),
+			CreatedAt: uint32(1655000000 + i),
 			Kind:      1,
 			Tags:      [][]string{{"t", "test"}},
 			Content:   "Message " + string(rune('A'+i)),
@@ -325,7 +325,7 @@ func TestScannerMultiPage(t *testing.T) {
 		event := &types.Event{
 			ID:        [32]byte{byte(i)},
 			Pubkey:    [32]byte{0x01},
-			CreatedAt: uint64(1655000000 + i),
+			CreatedAt: uint32(1655000000 + i),
 			Kind:      1,
 			Tags:      [][]string{},
 			Content:   ev.content,
